requirements.txt
# Backend Dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# AI/ML Dependencies
transformers==4.35.2
torch==2.1.1
sentence-transformers==2.2.2
langchain==0.0.335
chromadb==0.4.17
huggingface-hub==0.19.4

# Frontend Dependencies
streamlit==1.28.1
streamlit-option-menu==0.3.6
plotly==5.17.0

# Utility Dependencies
python-multipart==0.0.6
python-dotenv==1.0.0
pydantic==2.5.0
pandas==2.1.3
numpy==1.25.2
pillow==10.1.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4